[shim] Fix macOS build
https://github.com/dstackai/dstack/pull/2952 broke macOS because of
go-dcgm
