+++ runner/internal/shim/dcgm/wrapper_darwin.go
@@ -1,0 +1,9 @@
+//go:build darwin
+
+package dcgm
+
+import "errors"
+
+func NewDCGMWrapper(address string) (DCGMWrapperInterface, error) {
+	return nil, errors.New("macOS is not supported")
+}
